refactor(wave0d.1 #252): move agent/ pure-logic to daemon
Move electron/agent/read-default-model.ts (pure: only node:fs,
node:path + cross-package import of shared/claudePaths) into
packages/daemon/src/agent/. No current consumer in electron/, but the
re-export shim is kept for symmetry — orphan deletion can happen any
time.

[REFACTOR-ONLY] no behavior change.
